requirements.txt
aiofiles==24.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.2.1
pytest==7.4.3
requests==2.31.0
pyyaml==6.0.1
loguru==0.7.0
groq==0.18.0